+++ src/entities/chess_board.rs
@@ -1,162 +1,22 @@
-use std::collections::{BTreeMap, HashSet};
-use std::ops::Not;
+use std::collections::HashSet;
 
-use cfg_if::cfg_if;
-use leptos::{
-    create_rw_signal, RwSignal, Scope, SignalGetUntracked, SignalSet, SignalUpdate,
-    SignalWithUntracked,
-};
-use web_sys::WebSocket;
+use crate::entities::chess_board::enums::{PromotionKind, CastlePosition};
+
+use self::castle_rules::{fen_to_castle_rules, CastleRules, CastleOptions};
+use self::enums::{FenError, ChessBoardError, Move, MoveError};
+use self::passants::fen_to_passant;
+use self::stones::fen_to_stones;
+use self::turns::{fen_to_turn, Turn};
 
 use super::position::Position;
-use super::room::RoomStatus;
 use super::stone::{Color, Kind, Stone};
 
-pub fn fen_to_passant(field: &str) -> Result<Option<Position>, ()> {
-    if field == "-" {
-        return Ok(None);
-    }
-
-    let x = field.chars().nth(0).ok_or(())? as usize - 97;
-    let y = 8 - field.chars().nth(1).ok_or(())?.to_digit(10).ok_or(())? as usize;
-    Ok(Some(Position::new(x, y)))
-}
-
-pub fn fen_to_castle_rules(field: &str) -> Result<CastleRules, ()> {
-    let mut white_castle_options = CastleOptions::None;
-    let mut black_castle_options = CastleOptions::None;
-
-    for c in field.chars() {
-        match c {
-            'K' => {
-                white_castle_options = CastleOptions::KingSide;
-            }
-            'Q' => match white_castle_options {
-                CastleOptions::KingSide => {
-                    white_castle_options = CastleOptions::BothSides;
-                }
-                _ => {
-                    white_castle_options = CastleOptions::QueenSide;
-                }
-            },
-            'k' => {
-                black_castle_options = CastleOptions::KingSide;
-            }
-            'q' => match black_castle_options {
-                CastleOptions::KingSide => {
-                    black_castle_options = CastleOptions::BothSides;
-                }
-                _ => {
-                    black_castle_options = CastleOptions::QueenSide;
-                }
-            },
-            '-' => {
-                break;
-            }
-            _ => {
-                return Err(());
-            }
-        }
-    }
-
-    Ok(CastleRules {
-        white: white_castle_options,
-        black: black_castle_options,
-    })
-}
-
-pub fn fen_to_stones(field: &str) -> Result<[[Option<Stone>; 8]; 8], ()> {
-    const INIT: Option<Stone> = None;
-    const ROW: [Option<Stone>; 8] = [INIT; 8];
-    let mut stones = [ROW; 8];
-
-    for (y, row) in field.split("/").enumerate() {
-        if y > 7 {
-            return Err(());
-        }
-
-        let mut empty_squares = 0;
-
-        for (x, c) in row.chars().enumerate() {
-            if x > 7 {
-                return Err(());
-            }
-
-            if c.is_digit(10) {
-                let n = c.to_digit(10).ok_or(())?;
-                let x = x + empty_squares as usize;
-                empty_squares += n - 1;
-                for i in 0..n {
-                    let x = x + i as usize;
-                    if x > 7 {
-                        return Err(());
-                    }
-                    stones[y][x] = None;
-                }
-            } else {
-                let x = x + empty_squares as usize;
-                if x > 7 {
-                    return Err(());
-                }
-                stones[y][x] = Some(Stone::try_from(c)?);
-            }
-        }
-    }
-
-    Ok(stones)
-}
-
-pub fn fen_to_turn(field: &str) -> Result<Turn, ()> {
-    if field == "w" {
-        Ok(Turn::White)
-    } else if field == "b" {
-        Ok(Turn::Black)
-    } else {
-        Err(())
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum CastleOptions {
-    KingSide,
-    QueenSide,
-    BothSides,
-    None,
-}
-
-#[derive(Clone, Debug)]
-#[allow(dead_code)]
-pub struct CastleRules {
-    white: CastleOptions,
-    black: CastleOptions,
-}
-
-impl CastleRules {
-    pub fn white(&self) -> &CastleOptions {
-        &self.white
-    }
-
-    pub fn black(&self) -> &CastleOptions {
-        &self.black
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-pub enum Turn {
-    White,
-    Black,
-}
-
-impl Not for Turn {
-    type Output = Turn;
-
-    fn not(self) -> Self::Output {
-        match self {
-            Turn::White => Turn::Black,
-            Turn::Black => Turn::White,
-        }
-    }
-}
+pub mod castle_rules;
+pub mod enums;
+pub mod stones;
+pub mod turns;
+pub mod passants;
+pub mod signals;
 
 pub struct ChessBoardBuilder {
     fen: Option<String>,
@@ -203,42 +63,45 @@ impl ChessBoardBuilder {
         self
     }
 
-    pub fn build(self) -> Result<ChessBoard, ()> {
+    pub fn build(self) -> Result<ChessBoard, ChessBoardError> {
         let Some(fen) = self.fen else {
-            return Err(());
+            return Err(ChessBoardError::BuildError);
         };
 
         let Some(deleted_stones_str) = self.deleted_stones else {
-            return Err(());
+            return Err(ChessBoardError::BuildError);
         };
 
         let Some(is_white_view) = self.is_white_view else {
-            return Err(());
+            return Err(ChessBoardError::BuildError);
         };
 
         let Some(validation) = self.validation else {
-            return Err(());
+            return Err(ChessBoardError::BuildError);
         };
 
         let Some(sync) = self.sync else {
-            return Err(());
+            return Err(ChessBoardError::BuildError);
         };
 
         let fen_fields = fen.split(" ").collect::<Vec<&str>>();
+        if fen_fields.len() != 6 {
+            return Err(ChessBoardError::InvalidFen(FenError::InvalidFormat));
+        }
         let mut deleted_stones = Vec::new();
         for c in deleted_stones_str.chars() {
-            deleted_stones.push(Stone::try_from(c)?);
+            deleted_stones.push(Stone::try_from(c).map_err(|_| ChessBoardError::InvalidDeletedStones)?);
         }
 
         let mut chess_board = ChessBoard {
             fen: fen.to_string(),
-            stones: fen_to_stones(fen_fields[0])?,
+            stones: fen_to_stones(fen_fields[0]).map_err(|e| ChessBoardError::InvalidFen(e))?,
             treat_map: [[false; 8]; 8],
-            turn: fen_to_turn(fen_fields[1])?,
-            castle_rules: fen_to_castle_rules(fen_fields[2])?,
-            passant: fen_to_passant(fen_fields[3])?,
-            half_move_clock: fen_fields[4].parse::<i32>().map_err(|_| ())?,
-            full_move_clock: fen_fields[5].parse::<i32>().map_err(|_| ())?,
+            turn: fen_to_turn(fen_fields[1]).map_err(|e| ChessBoardError::InvalidFen(e))?,
+            castle_rules: fen_to_castle_rules(fen_fields[2]).map_err(|e| ChessBoardError::InvalidFen(e))?,
+            passant: fen_to_passant(fen_fields[3]).map_err(|e| ChessBoardError::InvalidFen(e))?,
+            half_move_clock: fen_fields[4].parse::<i32>().map_err(|_| ChessBoardError::InvalidFen(FenError::InvalidHalfMoveClock))?,
+            full_move_clock: fen_fields[5].parse::<i32>().map_err(|_| ChessBoardError::InvalidFen(FenError::InvalidFullMoveClock))?,
             deleted_stones,
             is_white_view,
             validation,
@@ -250,7 +113,7 @@ impl ChessBoardBuilder {
         }
 
         if chess_board.validation && !chess_board.valid_castle_rules() {
-            return Err(());
+            return Err(ChessBoardError::InvalidFen(FenError::InvalidCastleRules));
         }
 
         Ok(chess_board)
@@ -275,17 +138,17 @@ pub struct ChessBoard {
 
 #[allow(dead_code)]
 impl ChessBoard {
-    pub fn new(fen: &str) -> Result<Self, ()> {
+    pub fn new(fen: &str) -> Result<Self, ChessBoardError> {
         let fen_fields = fen.split(" ").collect::<Vec<&str>>();
         let mut chess_board = Self {
             fen: fen.to_string(),
-            stones: fen_to_stones(fen_fields[0])?,
+            stones: fen_to_stones(fen_fields[0]).map_err(|e| ChessBoardError::InvalidFen(e))?,
             treat_map: [[false; 8]; 8],
-            turn: fen_to_turn(fen_fields[1])?,
-            castle_rules: fen_to_castle_rules(fen_fields[2])?,
-            passant: fen_to_passant(fen_fields[3])?,
-            half_move_clock: fen_fields[4].parse::<i32>().map_err(|_| ())?,
-            full_move_clock: fen_fields[5].parse::<i32>().map_err(|_| ())?,
+            turn: fen_to_turn(fen_fields[1]).map_err(|e| ChessBoardError::InvalidFen(e))?,
+            castle_rules: fen_to_castle_rules(fen_fields[2]).map_err(|e| ChessBoardError::InvalidFen(e))?,
+            passant: fen_to_passant(fen_fields[3]).map_err(|e| ChessBoardError::InvalidFen(e))?,
+            half_move_clock: fen_fields[4].parse::<i32>().map_err(|_| ChessBoardError::InvalidFen(FenError::InvalidHalfMoveClock))?,
+            full_move_clock: fen_fields[5].parse::<i32>().map_err(|_| ChessBoardError::InvalidFen(FenError::InvalidFullMoveClock))?,
             deleted_stones: Vec::new(),
             is_white_view: true,
             validation: false,
@@ -295,7 +158,7 @@ impl ChessBoard {
             chess_board.sync_treat_map();
         }
         if chess_board.validation && !chess_board.valid_castle_rules() {
-            return Err(());
+            return Err(ChessBoardError::InvalidFen(FenError::InvalidCastleRules));
         }
         Ok(chess_board)
     }
@@ -317,7 +180,7 @@ impl ChessBoard {
     pub fn valid_castle_rules(&self) -> bool {
         match self.castle_rules.white() {
             CastleOptions::KingSide => {
-                if !self.stone_at_is(5, 7, Kind::King) {
+                if !self.stone_at_is(4, 7, Kind::King) {
                     return false;
                 }
 
@@ -326,7 +189,7 @@ impl ChessBoard {
                 }
             }
             CastleOptions::QueenSide => {
-                if !self.stone_at_is(5, 7, Kind::King) {
+                if !self.stone_at_is(4, 7, Kind::King) {
                     return false;
                 }
 
@@ -335,7 +198,7 @@ impl ChessBoard {
                 }
             }
             CastleOptions::BothSides => {
-                if !self.stone_at_is(5, 7, Kind::King) {
+                if !self.stone_at_is(4, 7, Kind::King) {
                     return false;
                 }
 
@@ -352,7 +215,7 @@ impl ChessBoard {
 
         match self.castle_rules.black() {
             CastleOptions::KingSide => {
-                if !self.stone_at_is(5, 0, Kind::King) {
+                if !self.stone_at_is(4, 0, Kind::King) {
                     return false;
                 }
 
@@ -361,7 +224,7 @@ impl ChessBoard {
                 }
             }
             CastleOptions::QueenSide => {
-                if !self.stone_at_is(5, 0, Kind::King) {
+                if !self.stone_at_is(4, 0, Kind::King) {
                     return false;
                 }
 
@@ -370,7 +233,7 @@ impl ChessBoard {
                 }
             }
             CastleOptions::BothSides => {
-                if !self.stone_at_is(5, 0, Kind::King) {
+                if !self.stone_at_is(4, 0, Kind::King) {
                     return false;
                 }
 
@@ -507,11 +370,11 @@ impl ChessBoard {
         }
     }
 
-    pub fn possible_moves(&self, position: &Position) -> Vec<Position> {
+    pub fn possible_moves(&self, position: &Position) -> HashSet<Position> {
         let Some(ref stone) = self.stone_at(position.x, position.y) else {
-            return Vec::new();
+            return HashSet::new();
         };
-        let mut moves = Vec::new();
+        let mut moves = HashSet::new();
         for possible_move in stone.possible_moves(position, &self) {
             let mut chess_board = self.clone();
             chess_board.validation = false;
@@ -523,7 +386,7 @@ impl ChessBoard {
             );
             chess_board.sync_treat_map();
             if !chess_board.is_in_check() {
-                moves.push(possible_move);
+                moves.insert(possible_move);
             }
         }
         moves
@@ -534,55 +397,192 @@ impl ChessBoard {
         piece: &str,
         from: Option<Position>,
         to: Option<Position>,
-    ) -> Result<(), ()> {
+    ) -> Result<Move, ChessBoardError> {
         if from == to {
-            return Ok(());
+            return Ok(Move::Normal);
         }
 
+        if self.validation {
+            let stone = piece.parse::<Stone>().map_err(|_| ChessBoardError::InvalidMove(MoveError::NoStoneFound))?;
+
+            match (stone.color(), self.turn) {
+                (Color::Light, Turn::White) | (Color::Dark, Turn::Black) => {}
+                _ => return Err(ChessBoardError::InvalidMove(MoveError::InvalidMove)),
+            }
+
+            if from.is_none() || to.is_none() {
+                return Err(ChessBoardError::InvalidMove(MoveError::InvalidMove));
+            }
+
+            let to = to.clone().unwrap();
+            let from = from.clone().unwrap();
+            let possible_moves = self.possible_moves(&from);
+            if !possible_moves.contains(&to) {
+                return Err(ChessBoardError::InvalidMove(MoveError::InvalidMove));
+            }
+        }
+
+        let Some(stone) = (match from.clone() {
+            None => {
+                if let Some(idx) = self
+                    .deleted_stones
+                    .iter()
+                    .position(|s| s.image_class() == piece)
+                    {
+                        Some(self.deleted_stones.remove(idx))
+                    } else {
+                        return Err(ChessBoardError::InvalidMove(MoveError::NoStoneFound));
+                    }
+            }
+            Some(from) => self.take_stone_at(from.x, from.y) 
+        }) else {
+            return Err(ChessBoardError::InvalidMove(MoveError::NoStoneFound));
+        };
+
+        let result;
+
         if from.is_none() {
-            if let Some(stone_idx) = self
-                .deleted_stones
-                .iter()
-                .position(|s| s.image_class() == piece)
-            {
-                let stone = self.deleted_stones.remove(stone_idx);
-                let to = to.unwrap();
-                let old_piece = self.take_stone_at(to.x, to.y);
-                self.stones[to.y as usize][to.x as usize] = Some(stone);
-                if let Some(old_piece) = old_piece {
-                    self.deleted_stones.push(old_piece);
-                }
-            } else {
-                return Err(());
-            };
-        } else if to.is_none() {
-            let from = from.unwrap();
-            let stone = self.take_stone_at(from.x, from.y);
-            if stone
-                .clone()
-                .map(|s| s.image_class() != piece)
-                .unwrap_or(true)
-            {
-                return Err(());
-            }
-            if let Some(stone) = stone {
-                self.deleted_stones.push(stone);
-            }
-        } else {
-            let from = from.unwrap();
             let to = to.unwrap();
-            let stone = self.take_stone_at(from.x, from.y);
-            if stone
-                .clone()
-                .map(|s| s.image_class() != piece)
-                .unwrap_or(true)
-            {
-                return Err(());
-            }
             let old_piece = self.take_stone_at(to.x, to.y);
-            self.stones[to.y as usize][to.x as usize] = stone;
+            self.stones[to.y as usize][to.x as usize] = Some(stone);
             if let Some(old_piece) = old_piece {
                 self.deleted_stones.push(old_piece);
+            }
+            result = Move::Normal;
+        } else if to.is_none() {
+            self.deleted_stones.push(stone);
+            result = Move::Normal;
+        } else {
+            let from = from.unwrap();
+            let to = to.unwrap(); 
+
+            let old_piece = self.take_stone_at(to.x, to.y);
+            self.stones[to.y as usize][to.x as usize] = Some(stone.clone());
+            if let Some(old_piece) = old_piece {
+                self.deleted_stones.push(old_piece);
+            }
+            let stone_str = stone.image_class();
+
+            log::debug!("Move {} from {:?} to {:?}", stone_str, from, to);
+
+            if self.validation {
+                if self.passant == Some(to.clone()) {
+                    let passant_pos = to.clone();
+                    let passant_stone = match stone.color() {
+                        Color::Light => {
+                            self.take_stone_at(passant_pos.x, passant_pos.y + 1)
+                        }
+                        Color::Dark => {
+                             self.take_stone_at(passant_pos.x, passant_pos.y - 1)
+                        }
+                    };
+                    self.deleted_stones.push(passant_stone.unwrap());
+                    self.passant = None;
+                    result = Move::Passant;
+                } else if stone_str == "lp" && from.y == 6 && to.y == 4 {
+                    self.passant = Some(Position::new(to.x, 5));
+                    result = Move::Normal;
+                } else if stone_str == "dp" && from.y == 1 && to.y == 3 {
+                    self.passant = Some(Position::new(to.x, 2));
+                    result = Move::Normal;
+                } else if stone_str == "lp" && to.y == 0 {
+                    self.passant = None;
+                    let new_stone: Stone = match (PromotionKind::Queen, stone.color()) {
+                        (PromotionKind::Queen, Color::Light) => 'Q'.try_into().unwrap(),
+                        (PromotionKind::Queen, Color::Dark) => 'q'.try_into().unwrap(),
+                        (PromotionKind::Rook, Color::Light) => 'R'.try_into().unwrap(),
+                        (PromotionKind::Rook, Color::Dark) => 'r'.try_into().unwrap(),
+                        (PromotionKind::Bishop, Color::Light) => 'B'.try_into().unwrap(),
+                        (PromotionKind::Bishop, Color::Dark) => 'b'.try_into().unwrap(),
+                        (PromotionKind::Knight, Color::Light) => 'N'.try_into().unwrap(),
+                        (PromotionKind::Knight, Color::Dark) => 'n'.try_into().unwrap(),
+                    };
+                    self.stones[to.y as usize][to.x as usize] = Some(new_stone);
+                    result = Move::Promotion(PromotionKind::Queen);
+                } else if stone_str == "dp" && to.y == 7 {
+                    self.passant = None;
+                    let new_stone: Stone = match (PromotionKind::Queen, stone.color()) {
+                        (PromotionKind::Queen, Color::Light) => 'Q'.try_into().unwrap(),
+                        (PromotionKind::Queen, Color::Dark) => 'q'.try_into().unwrap(),
+                        (PromotionKind::Rook, Color::Light) => 'R'.try_into().unwrap(),
+                        (PromotionKind::Rook, Color::Dark) => 'r'.try_into().unwrap(),
+                        (PromotionKind::Bishop, Color::Light) => 'B'.try_into().unwrap(),
+                        (PromotionKind::Bishop, Color::Dark) => 'b'.try_into().unwrap(),
+                        (PromotionKind::Knight, Color::Light) => 'N'.try_into().unwrap(),
+                        (PromotionKind::Knight, Color::Dark) => 'n'.try_into().unwrap(),
+                    };
+                    self.stones[to.y as usize][to.x as usize] = Some(new_stone);
+                    result = Move::Promotion(PromotionKind::Queen);
+                } else if stone_str == "lk" && from.x == 4 && from.y == 7 && to.x == 6 && to.y == 7 {
+                    let rook = self.take_stone_at(7, 7);
+                    self.stones[7][5] = rook;
+                    self.passant = None;
+                    self.castle_rules.white = CastleOptions::None;
+                    result = Move::Castle(CastlePosition::KingSide);
+                } else if stone_str == "lk" && from.x == 4 && from.y == 7 && to.x == 2 && to.y == 7 {
+                    let rook = self.take_stone_at(0, 7);
+                    self.stones[7][3] = rook;
+                    self.passant = None;
+                    self.castle_rules.white = CastleOptions::None;
+                    result = Move::Castle(CastlePosition::QueenSide);
+                } else if stone_str == "dk" && from.x == 4 && from.y == 0 && to.x == 6 && to.y == 0 {
+                    let rook = self.take_stone_at(7, 0);
+                    self.stones[0][5] = rook;
+                    self.passant = None;
+                    self.castle_rules.black = CastleOptions::None;
+                    result = Move::Castle(CastlePosition::KingSide);
+                } else if stone_str == "dk" && from.x == 4 && from.y == 0 && to.x == 2 && to.y == 0 {
+                    let rook = self.take_stone_at(0, 0);
+                    self.stones[0][3] = rook;
+                    self.passant = None;
+                    self.castle_rules.black = CastleOptions::None;
+                    result = Move::Castle(CastlePosition::QueenSide);
+                } else if stone_str == "lk" {
+                    self.passant = None;
+                    self.castle_rules.white = CastleOptions::None;
+                    result = Move::Normal;
+                } else if stone_str == "dk" {
+                    self.passant = None;
+                    self.castle_rules.black = CastleOptions::None;
+                    result = Move::Normal;
+                } else if stone_str == "lr" && from.x == 0 && from.y == 7 {
+                    self.passant = None;
+                    match self.castle_rules.white {
+                        CastleOptions::BothSides => self.castle_rules.white = CastleOptions::KingSide,
+                        CastleOptions::QueenSide => self.castle_rules.white = CastleOptions::None,
+                        _ => {} 
+                    }
+                    result = Move::Normal;
+                } else if stone_str == "lr" && from.x == 7 && from.y == 7 {
+                    self.passant = None;
+                    match self.castle_rules.white {
+                        CastleOptions::BothSides => self.castle_rules.white = CastleOptions::QueenSide,
+                        CastleOptions::KingSide => self.castle_rules.white = CastleOptions::None,
+                        _ => {} 
+                    }
+                    result = Move::Normal;
+                } else if stone_str == "dr" && from.x == 0 && from.y == 0 {
+                    self.passant = None;
+                    match self.castle_rules.black {
+                        CastleOptions::BothSides => self.castle_rules.black = CastleOptions::KingSide,
+                        CastleOptions::QueenSide => self.castle_rules.black = CastleOptions::None,
+                        _ => {} 
+                    }
+                    result = Move::Normal;
+                } else if stone_str == "dr" && from.x == 7 && from.y == 0 {
+                    self.passant = None;
+                    match self.castle_rules.black {
+                        CastleOptions::BothSides => self.castle_rules.black = CastleOptions::QueenSide,
+                        CastleOptions::KingSide => self.castle_rules.black = CastleOptions::None,
+                        _ => {} 
+                    }
+                    result = Move::Normal;
+                } else {
+                    self.passant = None;
+                    result = Move::Normal; 
+                }
+            } else {
+                result = Move::Normal;
             }
         }
 
@@ -592,7 +592,7 @@ impl ChessBoard {
             self.sync_treat_map();
         }
 
-        Ok(())
+        Ok(result)
     }
 
     pub fn sync_fen(&mut self) {
@@ -623,9 +623,15 @@ impl ChessBoard {
             new_fen.push_str(" b");
         }
         let fen_clone = self.fen.clone();
-        let fen_split = fen_clone.split(" ").skip(2);
-        let fen_rest = fen_split.collect::<Vec<&str>>().join(" ");
-        new_fen.push_str(&format!(" {}", fen_rest));
+        let mut fen_split = fen_clone.split(" ").skip(4);
+        let fen_castle_rules = self.castle_rules.to_string();
+        new_fen.push_str(&format!(" {}", fen_castle_rules));
+        let fen_passant = self.passant.as_ref().map(|pos| pos.to_string()).unwrap_or("-".to_string());
+        new_fen.push_str(&format!(" {}", fen_passant));
+        let fen_half_moves = fen_split.next().unwrap();
+        new_fen.push_str(&format!(" {}", fen_half_moves));
+        let fen_full_moves = fen_split.next().unwrap();
+        new_fen.push_str(&format!(" {}", fen_full_moves));
 
         self.fen = new_fen;
     }
@@ -644,484 +650,17 @@ impl ChessBoard {
     }
 }
 
-pub struct ChessBoardSignalsBuilder {
-    cx: Option<Scope>,
-    chess_board: Option<RwSignal<ChessBoard>>,
-    room_status: Option<RwSignal<Option<RoomStatus>>>,
-    chess_board_socket: Option<RwSignal<Option<WebSocket>>>,
-    stones_signals: Option<RwSignal<StonesSignals>>,
-    should_render: Option<RwSignal<bool>>,
-}
-
-impl ChessBoardSignalsBuilder {
-    pub fn new() -> Self {
-        Self {
-            cx: None,
-            chess_board: None,
-            room_status: None,
-            chess_board_socket: None,
-            stones_signals: None,
-            should_render: None,
-        }
-    }
-
-    pub fn cx(mut self, cx: Scope) -> Self {
-        self.cx = Some(cx);
-        self
-    }
-
-    pub fn chess_board(mut self, chess_board: RwSignal<ChessBoard>) -> Self {
-        self.chess_board = Some(chess_board);
-        self
-    }
-
-    pub fn room_status(mut self, room_status: RwSignal<Option<RoomStatus>>) -> Self {
-        self.room_status = Some(room_status);
-        self
-    }
-
-    pub fn chess_board_socket(mut self, chess_board_socket: RwSignal<Option<WebSocket>>) -> Self {
-        self.chess_board_socket = Some(chess_board_socket);
-        self
-    }
-
-    pub fn stones_signals(mut self, stones_signals: RwSignal<StonesSignals>) -> Self {
-        self.stones_signals = Some(stones_signals);
-        self
-    }
-
-    pub fn should_render(mut self, should_render: RwSignal<bool>) -> Self {
-        self.should_render = Some(should_render);
-        self
-    }
-
-    pub fn build(self) -> Result<ChessBoardSignals, ()> {
-        let Some(cx) = self.cx else {
-            return Err(());
-        };
-        let Some(chess_board) = self.chess_board else {
-            return Err(());
-        };
-        let Some(room_status) = self.room_status else {
-            return Err(());
-        };
-        let Some(chess_board_socket) = self.chess_board_socket else {
-            return Err(());
-        };
-        let Some(stones_signals) = self.stones_signals else {
-            return Err(());
-        };
-        let Some(should_render) = self.should_render else {
-            return Err(());
-        };
-
-        Ok(ChessBoardSignals {
-            cx,
-            chess_board,
-            room_status,
-            chess_board_socket,
-            stones_signals,
-            should_render,
-        })
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct StonesSignals {
-    board_stones: BTreeMap<String, RwSignal<StoneSignal>>,
-    deleted_stones_idx: usize,
-    deleted_stones: BTreeMap<usize, RwSignal<StoneSignal>>,
-}
-
-#[allow(dead_code)]
-impl StonesSignals {
-    pub fn new() -> Self {
-        Self {
-            board_stones: BTreeMap::new(),
-            deleted_stones: BTreeMap::new(),
-            deleted_stones_idx: 0,
-        }
-    }
-
-    pub fn get_board_stone(&self, key: &str) -> Option<RwSignal<StoneSignal>> {
-        self.board_stones.get(key).cloned()
-    }
-
-    pub fn get_deleted_stone(&self, idx: &usize) -> Option<RwSignal<StoneSignal>> {
-        self.deleted_stones.get(idx).cloned()
-    }
-
-    pub fn board_stones(&self) -> &BTreeMap<String, RwSignal<StoneSignal>> {
-        &self.board_stones
-    }
-
-    pub fn deleted_stones(&self) -> &BTreeMap<usize, RwSignal<StoneSignal>> {
-        &self.deleted_stones
-    }
-
-    pub fn add_board_stone(&mut self, cx: Scope, position: Position, stone: Stone) {
-        let stone_signal = StoneSignal::new(Some(position.clone()), stone);
-        let key = stone_signal.unique_key();
-        let stone_signal = create_rw_signal(cx, stone_signal);
-
-        self.board_stones.insert(key, stone_signal);
-    }
-
-    pub fn add_board_stone_signal(&mut self, key: String, stone_signal: RwSignal<StoneSignal>) {
-        self.board_stones.insert(key, stone_signal);
-    }
-
-    pub fn remove_board_stone(&mut self, key: String) -> Option<RwSignal<StoneSignal>> {
-        self.board_stones.remove(&key)
-    }
-
-    pub fn add_deleted_stone(&mut self, cx: Scope, stone: Stone) {
-        let stone_signal = StoneSignal::new_deleted(stone);
-        let stone_signal = create_rw_signal(cx, stone_signal);
-        self.deleted_stones
-            .insert(self.deleted_stones_idx, stone_signal);
-        self.deleted_stones_idx += 1;
-    }
-
-    pub fn add_deleted_stone_signal(&mut self, stone_signal: RwSignal<StoneSignal>) {
-        self.deleted_stones
-            .insert(self.deleted_stones_idx, stone_signal);
-        self.deleted_stones_idx += 1;
-    }
-
-    pub fn remove_deleted_stone(&mut self, idx: usize) -> Option<RwSignal<StoneSignal>> {
-        self.deleted_stones.remove(&idx)
-    }
-
-    pub fn clear_board_stones(&mut self) {
-        self.board_stones.clear();
-    }
-
-    pub fn clear_deleted_stones(&mut self) {
-        self.deleted_stones.clear();
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct StoneSignal {
-    position: Option<Position>,
-    stone: Stone,
-    dragging: bool,
-    deleted: bool,
-}
-
-#[allow(dead_code)]
-impl StoneSignal {
-    pub fn new(position: Option<Position>, stone: Stone) -> Self {
-        Self {
-            position,
-            stone,
-            dragging: false,
-            deleted: false,
-        }
-    }
-
-    pub fn new_deleted(stone: Stone) -> Self {
-        Self {
-            position: None,
-            stone,
-            dragging: false,
-            deleted: true,
-        }
-    }
-
-    pub fn is_deleted(&self) -> bool {
-        self.deleted
-    }
-
-    pub fn delete(&mut self) {
-        self.deleted = true;
-    }
-
-    pub fn restore(&mut self) {
-        self.deleted = false;
-    }
-
-    pub fn enable_dragging(&mut self) {
-        self.dragging = true;
-    }
-
-    pub fn disable_dragging(&mut self) {
-        self.dragging = false;
-    }
-
-    pub fn is_dragging(&self) -> bool {
-        self.dragging
-    }
-
-    pub fn set_position(&mut self, position: Option<Position>) {
-        self.position = position;
-    }
-
-    pub fn position(&self) -> Option<Position> {
-        self.position.clone()
-    }
-
-    pub fn stone(&self) -> Stone {
-        self.stone.clone()
-    }
-
-    pub fn unique_key(&self) -> String {
-        format!(
-            "{}_{}",
-            self.position
-                .clone()
-                .map(|p| p.to_string())
-                .unwrap_or("deleted".to_string()),
-            self.stone.char()
-        )
-    }
-}
-
-#[allow(dead_code)]
-#[derive(Copy, Clone)]
-pub struct ChessBoardSignals {
-    cx: Scope,
-    chess_board: RwSignal<ChessBoard>,
-    room_status: RwSignal<Option<RoomStatus>>,
-    chess_board_socket: RwSignal<Option<WebSocket>>,
-    stones_signals: RwSignal<StonesSignals>,
-    should_render: RwSignal<bool>,
-}
-
-#[allow(dead_code)]
-impl ChessBoardSignals {
-    pub fn cx(&self) -> Scope {
-        self.cx.clone()
-    }
-
-    pub fn socket(&self) -> RwSignal<Option<WebSocket>> {
-        self.chess_board_socket
-    }
-
-    pub fn room_status(&self) -> RwSignal<Option<RoomStatus>> {
-        self.room_status
-    }
-
-    pub fn chess_board(&self) -> RwSignal<ChessBoard> {
-        self.chess_board
-    }
-
-    pub fn stones_signals(&self) -> RwSignal<StonesSignals> {
-        self.stones_signals
-    }
-
-    pub fn should_render(&self) -> RwSignal<bool> {
-        self.should_render
-    }
-
-    pub fn move_piece(&self, piece: String, old_pos: String, new_pos: String) {
-        if old_pos == new_pos {
-            return;
-        }
-
-        let old_pos = match old_pos.as_str() {
-            "deleted" => None,
-            _ => Some(old_pos.parse::<Position>().unwrap()),
-        };
-        let new_pos = match new_pos.as_str() {
-            "deleted" => None,
-            _ => Some(new_pos.parse::<Position>().unwrap()),
-        };
-
-        let old_pos_clone = old_pos.clone();
-        let new_pos_clone = new_pos.clone();
-
-        let stones_signals = self.stones_signals().get_untracked();
-        let deleted_stones = stones_signals.deleted_stones();
-        let mut new_pos_stone: Option<Stone> = None;
-        if let Some(new_pos) = new_pos.clone() {
-            let x = new_pos.x;
-            let y = new_pos.y;
-            new_pos_stone = self
-                .chess_board()
-                .get_untracked()
-                .stone_at(x, y)
-                .map(|s| s.clone());
-        }
-
-        self.chess_board().update(|chessboard| {
-            let _ = chessboard.move_piece(&piece, old_pos_clone, new_pos_clone);
-        });
-
-        if old_pos.is_none() {
-            let new_pos = new_pos.expect("new_pos should be Some");
-
-            let stone_key = deleted_stones
-                .iter()
-                .find(|(_, stone)| *stone.get_untracked().stone().image_class() == piece);
-
-            if let Some(new_pos_stone) = new_pos_stone {
-                let stone_key =
-                    StoneSignal::new(Some(new_pos.clone()), new_pos_stone.clone()).unique_key();
-
-                if let Some(stone_signal) = self
-                    .stones_signals()
-                    .with_untracked(|ss| ss.get_board_stone(&stone_key))
-                {
-                    stone_signal.update(|ss| {
-                        ss.set_position(None);
-                        ss.delete();
-                    });
-                };
-
-                self.stones_signals().update(|stones| {
-                    if let Some(stone) = stones.remove_board_stone(stone_key) {
-                        stones.add_deleted_stone_signal(stone);
-                    };
-                });
-            }
-
-            let new_key = StoneSignal::new(
-                Some(new_pos.clone()),
-                piece
-                    .parse::<Stone>()
-                    .expect("piece_data should be a valid stone"),
-            )
-            .unique_key();
-
-            if let Some((key, stone_signal)) = stone_key.map(|sk| sk) {
-                stone_signal.update(|ss| {
-                    ss.set_position(Some(new_pos.clone()));
-                    ss.disable_dragging();
-                    ss.restore();
-                });
-
-                let key = key.clone();
-                self.stones_signals().update(|stones| {
-                    if let Some(stone) = stones.remove_deleted_stone(key) {
-                        stones.add_board_stone_signal(new_key, stone);
-                    };
-                });
-            }
-        } else if new_pos.is_none() {
-            let old_pos = old_pos.expect("old_pos should be Some");
-
-            let stone_signal_key = StoneSignal::new(
-                Some(old_pos),
-                piece
-                    .parse::<Stone>()
-                    .expect("piece_data should be a valid stone"),
-            )
-            .unique_key();
-
-            let mut stone_signal: Option<RwSignal<StoneSignal>> = None;
-
-            self.stones_signals().update(|stones| {
-                if let Some(stone) = stones.remove_board_stone(stone_signal_key) {
-                    stones.add_deleted_stone_signal(stone);
-                    stone_signal = Some(stone);
-                }
-            });
-
-            if let Some(stone_signal) = stone_signal {
-                stone_signal.update(|ss| {
-                    ss.set_position(None);
-                    ss.delete();
-                    ss.disable_dragging();
-                });
-            }
-        } else {
-            let old_pos = old_pos.expect("old_pos should be Some");
-            let new_pos = new_pos.expect("new_pos should be Some");
-
-            let stone_signal_key = StoneSignal::new(
-                Some(old_pos),
-                piece
-                    .parse::<Stone>()
-                    .expect("piece_data should be a valid stone"),
-            )
-            .unique_key();
-
-            let new_pos_stone_signal_key = new_pos_stone
-                .map(|stone| StoneSignal::new(Some(new_pos.clone()), stone.clone()).unique_key());
-
-            let mut stone_signal: Option<RwSignal<StoneSignal>> = None;
-            let mut deleted_stone_signal: Option<RwSignal<StoneSignal>> = None;
-            let new_key = StoneSignal::new(
-                Some(new_pos.clone()),
-                piece
-                    .parse::<Stone>()
-                    .expect("piece_data should be a valid stone"),
-            )
-            .unique_key();
-
-            if let Some(key) = new_pos_stone_signal_key.clone() {
-                if let Some(stone_signal) = self
-                    .stones_signals()
-                    .with_untracked(|ss| ss.get_board_stone(&key))
-                {
-                    stone_signal.update(|ss| {
-                        ss.set_position(None);
-                        ss.delete();
-                    });
-                };
-
-                self.stones_signals.update(|stones| {
-                    if let Some(stone) = stones.remove_board_stone(key) {
-                        stones.add_deleted_stone_signal(stone);
-                        deleted_stone_signal = Some(stone);
-                    };
-                });
-            }
-
-            self.stones_signals().update(|stones| {
-                if let Some(stone) = stones.remove_board_stone(stone_signal_key) {
-                    stones.add_board_stone_signal(new_key, stone);
-                    stone_signal = Some(stone);
-                }
-            });
-
-            if let Some(deleted_stone_signal) = deleted_stone_signal {
-                deleted_stone_signal.update(|ss| {
-                    ss.set_position(None);
-                    ss.disable_dragging();
-                });
-            }
-
-            if let Some(stone_signal) = stone_signal {
-                stone_signal.update(|ss| {
-                    ss.set_position(Some(new_pos.clone()));
-                    ss.disable_dragging();
-                });
-            }
-        }
-    }
-
-    #[allow(unused_variables)]
-    pub fn start_websocket(&self) {
-        if self.chess_board_socket.with_untracked(|ws| ws.is_some()) {
-            return;
-        }
-
-        cfg_if! {
-            if #[cfg(not(feature = "ssr"))] {
-                let ws = crate::client::websockets::chess_board::start_websocket(*self).ok();
-            } else {
-                let ws = None;
-            }
-        }
-
-        self.chess_board_socket.set(ws);
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
 
     #[test]
     fn test_chess_board() {
-        let fen = "3k1bnr/p3pppp/3b4/8/8/8/PPP2PPP/RNBQKBNR b KQk - 0 1";
-        let chess_board = ChessBoard::new(fen).unwrap();
+        let fen = "rnb1kbnr/ppp1p1pp/8/4Pp2/8/3P1N2/PP1P1PPP/RNBQK2R w KQkq f6 0 1"; 
+        let chess_board = ChessBoardBuilder::new().fen(fen).validation(true).sync(true).build().unwrap();
 
         assert_eq!(
-            Vec::<Position>::new(),
+            HashSet::<Position>::new(),
             chess_board.possible_moves(&Position::new(3, 2))
         );
     }
